added comments in hardware added code for duplicate POST required

+++ resources/hardware.py
@@ -5,7 +5,7 @@ from datetime import datetime
 from sqlalchemy.orm.exc import NoResultFound
 from common.json_schema import Hardware_Schema
 from common.utils import headers,is_logged_in,has_admin_privileges
-from common.utils import bad_request,unauthorized,forbidden,not_found,internal_server_error,unprocessable_entity
+from common.utils import bad_request,unauthorized,forbidden,not_found,internal_server_error,unprocessable_entity,conflict
 
 class Hardware_RUD(Resource):
     """
@@ -15,6 +15,8 @@ class Hardware_RUD(Resource):
         """
         GET the hardware details based on specific hardware_id
         """
+        #using get instead of query and it is marginally faster than filter
+        #check for multiple entries need to be done at POST and not during GET or PUT or DELETE
         hardware_item = g.session.query(g.Base.classes.hardware).get(hardware_id)
         if hardware_item:
             ret = Hardware_Schema().dump(hardware_item).data
@@ -72,11 +74,12 @@ class Hardware_RUD(Resource):
         if user_status in ["director","organizer"]:
             try:
                 #this makes sure that at least one hardware matches hardware id
-                g.session.query(g.Base.classes.hardware).filter(g.Base.classes.hardware.id == hardware_id).first()
-                g.session.query(g.Base.classes.hardware).filter(g.Base.classes.hardware.id == hardware_id).delete()
-                return ("",204,headers)
-            except NoResultFound:
-                return (not_found,404,headers)
+                hardware_to_delete = g.session.query(g.Base.classes.hardware).get(hardware_id)
+                if hardware_to_delete:
+                    g.session.query(g.Base.classes.hardware).filter(g.Base.classes.hardware.id == hardware_id).delete()
+                    return ("",204,headers)
+                else:
+                    return (not_found,404,headers)
             except:
                 return (internal_server_error,500,headers)
         else:
@@ -86,17 +89,6 @@ class Hardware_CR(Resource):
     """
     To create new hardware using POST and read all hardware items
     """
-    def get(self):
-        """
-        GET all the announcements at a time.
-        """
-        try:
-            all_hardware = g.session.query(g.Base.classes.hardware).all()
-            ret = Hardware_Schema(many = True).dump(all_hardware).data
-            return (ret,200,headers)
-        except:
-            return (internal_server_error,500,headers)
-
     def post(self):
         """
         Create new hardware. Required data: item,lender, quantity
@@ -118,6 +110,10 @@ class Hardware_CR(Resource):
         if user_status == "not_logged_in":
             return (unauthorized,401,headers)
 
+        exist_check = g.session.query(exists().where(and_(g.Base.classes.hardware.item == data["item"],g.Base.classes.hardware.lender == data["lender"])).scalar()
+        if exist_check:
+            return (conflict,409,headers)
+
         if user_status in ["director","organizer"]:
             Hardware = g.Base.classes.hardware
             try:
@@ -131,11 +127,21 @@ class Hardware_CR(Resource):
                 g.session.add(new_hardware)
                 g.session.commit()
                 ret = g.session.query(g.Base.classes.hardware).filter(g.Base.classes.hardware.item == data["item"]).one()
-                ret = Hardware_Schema().dump(ret).data
-                return (ret,201 ,headers)
+                return (Hardware_Schema().dump(ret).data,201 ,headers)
             except Exception as err:
                 print(type(err))
                 print(err)
                 return (internal_server_error,500,headers)
         else:
             return(forbidden,403,headers)
+
+    def get(self):
+        """
+        GET all the announcements at a time.
+        """
+        try:
+            all_hardware = g.session.query(g.Base.classes.hardware).all()
+            ret = Hardware_Schema(many = True).dump(all_hardware).data
+            return (ret,200,headers)
+        except:
+            return (internal_server_error,500,headers)
